% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{g}
\alias{g}
\title{Title}
\format{
An object of class \code{data.frame} with 113 rows and 102 columns.
}
\usage{
g
}
\arguments{
\item{data}{A data frame}

\item{col}{A variable in the data frame}

\item{value}{The value of the variable you want to create a subset of.}
}
\value{
A subset of the data frame
}
\description{
my_select selects values from a specified variable in a data frame
}
\examples{
# Pass the data frame you want to create a subset of. The variable of interest and the value you want to select. Here we are creating a subset of data frame f gathering observations based from feather material.
}
\keyword{datasets}
